+++ src/lib.rs
@@ -81,6 +81,7 @@ pub mod negotiate;
 pub mod network_client;
 pub mod ntlm;
 pub mod pku2u;
+mod pk_init;
 #[cfg(windows)]
 pub mod winapi;
